Move the state-aware backends to engine.
