+++ src/core/mxc_engine/src/state_aware.rs
@@ -1,0 +1,43 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+//! State-aware lifecycle dispatch.
+//!
+//! The single home for resolving a parsed state-aware request to its backend
+//! and driving the per-phase flow. It centralizes the backend-specific
+//! construction that previously lived inline in `wxc-exec` so the binary can
+//! shrink to a thin CLI shell.
+//!
+//! Backends whose `StatefulSandboxBackend` impl lives in a `backends/*` crate
+//! (which depends on `wxc_common`, so the construction can't live inside
+//! `wxc_common` without a cycle) are constructed here — the engine already
+//! depends on those crates. Anything without a state-aware impl falls back to
+//! [`wxc_common::state_aware_dispatch::run_state_aware`], which surfaces the
+//! `unsupported_phase` envelope.
+
+use wxc_common::mxc_error::MxcError;
+use wxc_common::state_aware_dispatch::{
+    resolve_backend, run_state_aware as run_state_aware_fallback, DispatchOutcome,
+};
+use wxc_common::state_aware_request::ParsedStateAwareRequest;
+
+/// Resolve `parsed`'s backend and run the requested state-aware phase.
+///
+/// On envelope phases this returns [`DispatchOutcome::Envelope`]; on the exec
+/// phase it streams output live and returns
+/// [`DispatchOutcome::ExecCompleted`]. Dispatch failures return an
+/// [`MxcError`] the caller renders as a JSON error envelope.
+pub fn run_state_aware(
+    parsed: ParsedStateAwareRequest,
+    dry_run: bool,
+) -> Result<DispatchOutcome, MxcError> {
+    let backend = resolve_backend(&parsed)?;
+    match backend {
+        #[cfg(all(target_os = "windows", feature = "isolation_session"))]
+        wxc_common::models::ContainmentBackend::IsolationSession => {
+            let mut runner = isolation_session_common::IsolationSessionRunner::new();
+            wxc_common::state_aware_dispatch::dispatch_state_aware(&mut runner, parsed, dry_run)
+        }
+        _ => run_state_aware_fallback(parsed, dry_run),
+    }
+}
